requirements.txt
typing_extensions; python_version=="3.10"
absl-py==1.4.0
array-record==0.2.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dm-tree==0.1.8
etils==1.3.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
jax==0.4.11
jaxlib==0.4.11
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-collections==0.1.1
ml-dtypes==0.2.0
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
packaging==23.1
pgmax==0.6.0
Pillow==9.5.0
promise==2.3
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.5
wrapt==1.14.1
zipp==3.15.0
